Address review: cache datacell queryset, move formatCellValue to utils, improve banner copy
- Cache permission-filtered datacell queryset on ExtractType instance so
  resolve_full_datacell_list and resolve_datacell_count share one queryset
  build per request instead of duplicating the DB round-trip.
- Move formatCellValue from inline definition in ExtractGridEmbed.tsx to
  frontend/src/utils/formatters.ts per project convention for shared helpers.
- Clarify the combined truncation banner message to read "Showing N of M
  documents (payload bounded to X of Y total cells)" instead of the harder
  to parse previous phrasing.
- Add null guard on fullDatacellList access for defensive runtime safety.
- Add comment in GraphQL query noting offset is accepted but not yet used.
